# BE_FARMX/openapi.yaml
openapi: 3.0.0
info:
  title: FarmX API
  description: API for Farm/Pond/Scan Management System
  version: 1.0.0
  contact:
    name: FarmX Team
servers:
  - url: http://localhost:4000/api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [user, admin, expert]
        subLevel:
          type: integer
        ownedProducts:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users (admin only) with search and filtering
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: isActive
          schema:
            type: boolean
        - in: query
          name: role
          schema:
            type: string
            enum: [user, admin, expert]
        - in: query
          name: search
          schema:
            type: string
            description: Search by username, email, or fullName
        - in: query
          name: subLevel
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  totalPages:
                    type: integer
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user information (own profile or admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User information
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: "#/components/schemas/User"
    put:
      tags:
        - Users
      summary: Update user information (own profile or admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                fullName:
                  type: string
                phone:
                  type: string
                role:
                  type: string
                  enum: [user, admin, expert]
                  description: Only admin can change role
      responses:
        "200":
          description: User updated
    delete:
      tags:
        - Users
      summary: Delete a user (admin only or own account)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
  /users/{id}/subscription:
    post:
      tags:
        - Users
      summary: Update user subscription level
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subLevel
              properties:
                subLevel:
                  type: integer
                  minimum: 0
      responses:
        "200":
          description: Subscription updated